requirements.txt
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
boto3==1.35.57
botocore==1.35.57
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.11.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
emails==0.6
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
identify==2.6.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
narwhals==1.13.3
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
premailer==3.10.0
protobuf==5.28.3
psycopg==3.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruff==0.6.7
s3transfer==0.10.3
sentry-sdk==2.18.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
streamlit==1.40.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==14.0
